# api/GlobalTickets.Web/Models/Api/BasketLineForUpdate.cs
using System;
using System.ComponentModel.DataAnnotations;

namespace GloboTickets.Web.Models.Api
{
    public class BasketLineForUpdate
    {
        [Required]
        public Guid LineId { get; set; }
        [Required]
        public int TicketAmount { get; set; }
    }
}