# api/swagger2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GlobalTickets.Services.ShoppingBasket",
    "version": "v1"
  },
  "paths": {
    "/api/baskets/{basketId}/basketlines": {
      "get": {
        "tags": [
          "BasketLines"
        ],
        "parameters": [
          {
            "name": "basketId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BasketLine"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BasketLine"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BasketLine"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "BasketLines"
        ],
        "parameters": [
          {
            "name": "basketId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BasketLineForCreation"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BasketLineForCreation"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BasketLineForCreation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BasketLine"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasketLine"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasketLine"
                }
              }
            }
          }
        }
      }
    },
    "/api/baskets/{basketId}/basketlines/{basketLineId}": {
      "get": {
        "tags": [
          "BasketLines"
        ],
        "operationId": "GetBasketLine",
        "parameters": [
          {
            "name": "basketId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "basketLineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BasketLine"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasketLine"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasketLine"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "BasketLines"
        ],
        "parameters": [
          {
            "name": "basketId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "basketLineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BasketLineForUpdate"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BasketLineForUpdate"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BasketLineForUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BasketLine"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasketLine"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasketLine"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "BasketLines"
        ],
        "parameters": [
          {
            "name": "basketId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "basketLineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/baskets/{basketId}": {
      "get": {
        "tags": [
          "Baskets"
        ],
        "operationId": "GetBasket",
        "parameters": [
          {
            "name": "basketId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Basket"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Basket"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Basket"
                }
              }
            }
          }
        }
      }
    },
    "/api/baskets": {
      "post": {
        "tags": [
          "Baskets"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BasketForCreation"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BasketForCreation"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BasketForCreation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Basket"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Basket"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Basket"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Event": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "BasketLine": {
        "type": "object",
        "properties": {
          "basketLineId": {
            "type": "string",
            "format": "uuid"
          },
          "basketId": {
            "type": "string",
            "format": "uuid"
          },
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "price": {
            "type": "integer",
            "format": "int32"
          },
          "ticketAmount": {
            "type": "integer",
            "format": "int32"
          },
          "event": {
            "$ref": "#/components/schemas/Event"
          }
        },
        "additionalProperties": false
      },
      "BasketLineForCreation": {
        "required": [
          "eventId",
          "price",
          "ticketAmount"
        ],
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "price": {
            "type": "integer",
            "format": "int32"
          },
          "ticketAmount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "BasketLineForUpdate": {
        "required": [
          "ticketAmount"
        ],
        "type": "object",
        "properties": {
          "ticketAmount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "Basket": {
        "type": "object",
        "properties": {
          "basketId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "numberOfItems": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "BasketForCreation": {
        "required": [
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      }
    }
  }
}